hibak javitva

+++ main.cpp
@@ -15,7 +15,7 @@ int main()
         std::cout << "Ertek:" << b[i] <<std::endl;		//i output missing
     }
     std::cout << "Atlag szamitasa: " << std::endl;
-    float atlag;										//should be double
+    double atlag;										//should be double
     for (int i = 0; i < N_ELEMENTS, i++)
     {
         atlag += b[i]
